chore: add config system, fix bug of naive quant

+++ gact/gact/jpeg_processor.py
@@ -5,8 +5,8 @@ class JPEGProcessor(torch.nn.Module):
   def __init__(self, quality=75):
     super(JPEGProcessor, self).__init__()
     self.quality = quality
-    self.quant_matrix = get_dqf_matrix(quality, flatten=False).to('cuda')
-    self.dct_base = get_dct_matrix(8).to('cuda')
+    self.quant_matrix = get_dqf_matrix(quality, flatten=False).to(torch.float16).to('cuda')
+    self.dct_base = get_dct_matrix(8).to(torch.float16).to('cuda')
 
   def forward(self, x):
     # '''
